refactor: order 타입 파일 분리
