+++ src/types/types.ts
@@ -11,15 +11,6 @@ export type User = {
   avatarUrl?: string;
 };
 
-export type Order = {
-  date: string;
-  orderNumber: string;
-  products: { name: string; image: string; option: string; count: number }[];
-  amount: string;
-  status: string;
-  statusType: "success" | "processing" | "canceled" | "done";
-};
-
 export type Review = {
   id: string;
   product: { name: string; image: string; date: string };
